Move AST builder helpers into parse utils
