+++ packages/orga/src/parse/planning.ts
@@ -1,5 +1,6 @@
 import { Lexer } from '../tokenize'
 import { Planning } from '../types'
+import * as ast from './utils';
 
 export default (lexer: Lexer): Planning[] => {
 
@@ -12,15 +13,12 @@ export default (lexer: Lexer): Planning[] => {
     const timestamp = peek(1)
     if (!keyword || keyword.type !== 'planning.keyword') return
     if (!timestamp || timestamp.type !== 'planning.timestamp') return
-    const planning: Planning = {
-      type: 'planning',
-      keyword: keyword.value,
-      timestamp: timestamp.value,
+    const planning = ast.planning(keyword.value, timestamp.value, {
       position: {
         start: keyword.position.start,
         end: timestamp.position.end,
-      }
-    }
+      },
+    });
     eat()
     eat()
